requirements.txt
@"
Flask==3.0.0
Flask-MySQLdb==2.0.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
python-dotenv==1.0.0
marshmallow==3.20.1
mysql-connector-python==8.2.0
"@ | Out-File -FilePath "requirements.txt" -Encoding UTF8
